Revert "No funciona agregar insumo al comprar boleto"
This reverts commit f11dac712595474cc4c919106813497382166c74.

+++ models/viaje.py
@@ -9,7 +9,7 @@ class Viaje(db.Model):
     fecha = db.Column(db.Date)
     horaSalida = db.Column (db.Time())
     horaLlegada = db.Column (db.Time())
-    precio = db.Column(db.Integer)
+    precio = db.Column(db.Float)
     estado = db.Column(db.Integer)
     asientos = db.Column(db.Integer)
     enabled = db.Column(db.Integer)
